Switch to loop for reading, and single process.
